requirements.txt
pypdf
python-dotenv
transformers
einops
accelerate
langchain
bitsandbytes
sentence_transformers
llama-index
streamlit
torch==2.0.0
torchvision==0.15.1
llama-cpp-python
